Fix FTBFS due to Python2, and remove GTK3-based BuildRequires
Replace references to /usr/bin/python with /usr/bin/python2.

inkscape-0.92.3 uses GTK3 only if the WITH_GTK3_EXPERIMENTAL flag is
enabled during the build. Otherwise it uses GTK2. The Fedora package
currently doesn't use the GTK3 build, but the GTK2 build. Since it's
not possible to mix GTK2 and GTK3 in the same process, none of the
GTK3-based BuildRequires are necessary.

This reverts commit 9883cce40095bca04452 and brings back the gtkspell
requirement. gtkspell3 is a GTK3 library that's not used in the GTK2
build, anyway, and it was only working so far because the gtkspell
requirement is optional in the GTK2 build. Since gtkspell is no longer
orphaned there's no problem in bringing it back.

https://bugzilla.redhat.com/show_bug.cgi?id=1464487
https://bugzilla.redhat.com/show_bug.cgi?id=1604371